src/0061_rotate_list/rotate_list.go
package solution

import . "github.com/nimrc/leetcode/pkg/linkedlist"

/*
 * @lc app=leetcode id=61 lang=golang
 *
 * [61] Rotate List
 *
 * https://leetcode.com/problems/rotate-list/description/
 *
 * algorithms
 * Medium (31.96%)
 * Likes:    2286
 * Dislikes: 1150
 * Total Accepted:    371.1K
 * Total Submissions: 1.2M
 * Testcase Example:  '[1,2,3,4,5]\n2'
 *
 * Given the head of a linked list, rotate the list to the right by k
 * places.
 *
 *
 * Example 1:
 *
 *
 * Input: head = [1,2,3,4,5], k = 2
 * Output: [4,5,1,2,3]
 *
 *
 * Example 2:
 *
 *
 * Input: head = [0,1,2], k = 4
 * Output: [2,0,1]
 *
 *
 *
 * Constraints:
 *
 *
 * The number of nodes in the list is in the range [0, 500].
 * -100 <= Node.val <= 100
 * 0 <= k <= 2 * 10^9
 *
 *
 */

// @lc code=start
/**
 * Definition for singly-linked list.
 * type ListNode struct {
 *     Val int
 *     Next *ListNode
 * }
 */
func rotateRight(head *ListNode, k int) *ListNode {
	if k == 0 || head == nil || head.Next == nil {
		return head
	}

	size := 1
	iter := head

	for iter.Next != nil {
		iter = iter.Next
		size++
	}

	n := size - k%size

	if n == size {
		return head
	}

	iter.Next = head // 头尾相连

	for n > 0 {
		n--

		iter = iter.Next
	}
	head = iter.Next
	iter.Next = nil

	return head
}
